Specify allowed blocks for Product Template to core and woocommerce only
